db: restructure endpoint queries
