+++ storage/sqlite/db.go
@@ -81,27 +81,6 @@ func (d *SqliteBackend) GetChains() ([]types.Chain, error) {
 	return chains, nil
 }
 
-func (d *SqliteBackend) GetEnabledEndpoints(chainID int) ([]types.Endpoint, error) {
-	rows, err := d.db.Query("SELECT id, chain_id, host, port, protocol, username, password, enabled FROM endpoints WHERE chain_id = ? AND enabled = 1;", chainID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query endpoints: %w", err)
-	}
-
-	defer rows.Close()
-
-	var endpoints []types.Endpoint
-
-	for rows.Next() {
-		var endpoint types.Endpoint
-		if err := rows.Scan(&endpoint.ID, &endpoint.ChainID, &endpoint.Host, &endpoint.Port, &endpoint.Protocol, &endpoint.Username, &endpoint.Password, &endpoint.Enabled); err != nil {
-			return nil, fmt.Errorf("failed to scan endpoint: %w", err)
-		}
-		endpoints = append(endpoints, endpoint)
-	}
-
-	return endpoints, nil
-}
-
 func (d *SqliteBackend) UpsertChainTip(tip types.ChainTip) error {
 	sqlStmt := `INSERT INTO chaintips (chain_id, endpoint_id, height, hash, branchlen, status) VALUES (?, ?, ?, ?, ?, ?)
 				ON CONFLICT (height, hash, endpoint_id, chain_id) DO UPDATE SET branchlen = excluded.branchlen, status = excluded.status;`
